move mockPageStore into shared lib
