+++ packages/webapp/src/lib/__mocks__/stores.ts
@@ -3,24 +3,6 @@ import { type Config } from '@wagmi/core';
 import { mockWeb3Config } from './mockWeb3Config';
 import type { AppKit } from '@reown/appkit';
 
-const initialPageState = {
-	data: {
-		stores: { settings: {} },
-		dotrain: 'some dotrain content',
-		deployment: { key: 'deploy-key' },
-		strategyDetail: {}
-	},
-	url: new URL('http://localhost:3000/deploy'),
-	params: {},
-	form: {},
-	status: 200,
-	error: null,
-	route: {
-		id: null
-	}
-};
-
-const mockPageWritable = writable<typeof initialPageState>(initialPageState);
 const mockSignerAddressWritable = writable<string>('');
 const mockChainIdWritable = writable<number>(0);
 const mockConnectedWritable = writable<boolean>(false);
@@ -57,14 +39,3 @@ export const mockAppKitModalStore = {
 	mockSetSubscribeValue: (value: AppKit): void => mockAppKitModalWritable.set(value)
 };
 
-export const mockPageStore = {
-	subscribe: mockPageWritable.subscribe,
-	set: mockPageWritable.set,
-	mockSetSubscribeValue: (newValue: Partial<typeof initialPageState>): void => {
-		mockPageWritable.update((currentValue) => ({
-			...currentValue,
-			...newValue
-		}));
-	},
-	reset: () => mockPageWritable.set(initialPageState)
-};
